+++ Task_1_1_3/src/main/java/ru/nsu/pereverzev/Number.java
@@ -1,14 +1,14 @@
 package ru.nsu.pereverzev;
 
 public class Number extends Expression {
-    int value;
+    double value;
 
-    Number(int val) {
+    Number(double val) {
         value = val;
     }
 
     @Override
-    public int eval(String varsAsgn) {
+    public double eval(String varsAsgn) {
         return value;
     }
